--sample section Unit add,edit,delete,getAll

+++ src/main/java/com/square/Inventory/Management/System/Entity/SampleSectionUnit.java
@@ -13,7 +13,7 @@ import javax.persistence.*;
 public class SampleSectionUnit {
     @Id
     @Column(name = "ssuID")
-    private Integer ssuID;
+    private Long ssuID;
 
     @Column(name = "ssuName")
     private String ssuName;
